Make Deletion Logic more succinct

+++ Spindler/View Pages/Book Pages/BookDetailPage.xaml.cs
@@ -59,11 +59,7 @@ public partial class BookDetailPage : ContentPage, IQueryAttributable
 
     private async void DeleteButton_clicked(object sender, EventArgs e)
     {
-        if (!await DisplayAlert("Warning!", "Are you sure you want to delete this book?", "Yes", "No"))
-        {
-            return;
-        }
-        if (Book!.Id > 0)
+        if (Book!.Id > 0 && await DisplayAlert("Warning!", "Are you sure you want to delete this book?", "Yes", "No"))
         {
             await App.Database.DeleteItemAsync(Book);
         }
